Update arity of credit in adyen_common
Credit is called from spree with 4 args when it's possible to create a
payment profile, which is set to false explicitly in adyen_encrypted.rb - so it would never be
possible to call credit on payments taken with adyen encrypted without
errors. This is required in order to create refunds.