Fix transaction_id logging
Voided transactions were setting the payment transaction ID to nil,
effectively erasing the link between the original payment transaction ID
and the void transaction. This commit ensures that the *original*
payment transaction ID is preserved in the database. This does NOT do
anything to the transaction ID for the actual void transaction. That
transaction ID is presently being discarded.
Also adds code to log the refund's transaction ID on callbacks.

Addendum (Dylan): I removed AdyenPaymentCancellationResponse class that
is created at time of void as I could not figure out what it provided.