Added an uniqueness constraint to psp_reference
From Adyen's docs
> Adyen's unique reference associated with the transaction/the request.
This value is globally unique; quote it when communicating with us about this request.
https://docs.adyen.com/display/TD/Notification+fields

There is no reason not to enfoce uniqueness on this field, however I
explicitly want the error to be raised at the database should it ever be
violated (user error?) as a validation would cause silent failure and
loss of data. Atleast if the error is raised in the db the error output
would be visible in logs and Adyen would not receive an acknowledgement
for the notification, causing them to resend the notification once the
problem is corrected.