Added spec for the case that payment isn't authed
This control flow was untested, as seen in the test coverage.This
section should probably be changed, if the result is pending we
might get a response later in the callback saying the order is ready to
be captured. It's probably safe(?) From the docs: When authResult equals
PENDING, ERROR or CANCELLED, the pspReference may not yet be known;
therefore, it may be empty or not included.