Remove conditional block
As far as i can see, this doesn't work and never did. There's no tests
for this control block, and in development it silently fails - the first
time the notification is logged but when
@notification.original_reference is called it triggers an undefined
method exception. The exception causes the controller to never response
to adyen, which then sends the notification again - which is caught in
the duplicate record catch, allowing it to ack adyen.