Fix test_app generation
So unfortuanetly a lot of the stuff in here is blindly copied piece meal
from other gems. The structe of this gem wasn't good to begin with and
as a result a number of things like test_app generation didn't function
normally. This makes generating test apps **mostly** work, with the
exception that railties:install:migrations still needs to be run
manually to get things like solidus auth devise to work.