Added a link from payment to the CA area on Adyen.
Added a link that shows up on the payment page where currently
information about the source, and the history of notifications from
Adyen are displayed. The link has a style of .button.fa because Spree.
For some reason the there is a hardcoded style that changes the size of
the button when it has .fa applied to it...