Deleted everything not related to HPP
I'm unable to guarentee that any of other payment methods other than hpp
work. In addition they require you to be pci compliant which we, and our
clients our not. For these reasons I've elected to remove all code
related to it to try and minimize some of the signal to noise ratio. In
the future should we need this code we can just restore it at that time.