New means of processing notifications
Wrote new lib file call NotificationProcessing to handle notifications.
This extends the behavior of `handle!` that was originally in
AdyenNotification. We now actually use the `processed` field on
notifications when we do something with the notification!

done:
+ auto-capture authorizations complete the payment
+ capture completes the payment
+ failures fail the payment
missing:
- refunds
- void
- failed captures
- cancellations

As it is right now this lets us accept auto-capture payments. Next step
is to update the available options on the payment source so manual
captures work.