Added support back for manual captures
Because adyen payment modifications are delayed we don't actually know
if the request succeeded after doing it.

For that reason we can't use the standard capture! and friends as they
change the payment state and would result is false positives (payment
could fail after capture).

This is likely a bit fragile, but good enough for now.

When an AUTHORISATION event comes in it includes a list of modification
operations that may be preformed on it. We're going to use the name of
the event, prefixed with `adyen_hpp_` for the `fire` action in the
soldius backend.

This includes the implementation for adyen_hpp_capture, which is
essentially the first half of Spree::Payment::Processing#capture!. It
logs the event, marks the payment as processing and dispatches the
request for modification.

Once the capture event comes back, the payment is completed and the
captured amount is added as a `payment.capture_event`.

Captures are processed in the NotificationProcessing, added a few commits back, when the old capture! method was called it would register as a double charge once the adyen event was received.