Test to expose bug with early notifications
When process_outstanding is called, if there is already a authorization notification for a bank payment it completes the payment (from checkout -> completed states) - but this is prior to the order completing.

The order statemachine has a prerequisite transition event from `payment` -> `completed` called `process_paymnets_before_complete` which will call authorize on all payment in checkout state, or purchase if autocapture is enabled.

The problem arises when process_payment prevents transition to completed when there are no payments in checkout state