Fix issue with autocapture only methods
There was a bug where auto capture only payment methods were causing
completed orders to have a nil payment and shipping state.

Order is explicitly defined here because as of writing the Order ->
Payments association does not have the inverse of defined when we call
`current_order.complete` below payment.order will still refer to a
previous state of the record.

If the payment is auto captured only then the payment will completed in
`process_outstanding!`, and because Payment calls .order.update_totals
after save the order is saved with its previous values, causing
payment_state and shipment_state to revert to nil.