Fix erroring notifications
In the event that a notification cannot be processed we need to still
save the notification and acknoweldge it - otherwise Adyen will
continue to notify us about the event and it will continue to error it.

We cannot use an `ensure` in the controller action because the render
actually completes after the action. Doing so still results in a 500.

For this reason we need to save the notification on the first attempt,
let the controller error, and then on the second attempt from Adyen
return 200 [accepted], as the notification has already been saved.