Rename payment methods
Rename adyen_hpp_$action! methods to override capture! and friends.

This is required because order.cancel! calls .cancel on all payments so
to have it call our custom cancel logic we need to override it.

For everything that isn't a hpp payment we call the original
implementation of the method, with the exception of `credit!` which does
not have an implementation on all other payments.