Don't attempt to cancel payments don't support it
Attempting to cancel a payment that only supports manual refunds will
not contact the api.

Instead it will keep the payment in the completed state, which in turns
causes the order-payment-state to be 'credit owed'. A log entry in
created on the payment that has a message about the the manual refund.