Create payment on auth notification
We seem to have a weird situation where some payment methods may or may
not actually redirect the user back to confirmation page.

Previously the payment was only created on the payment page, as a
defensive mechanism we now create a payment and a source with whatever
we have available when we receive the auth notification IF the payment
doesn't already exist.