Prevent race condition between notification and redirect
Prior to this on the payment `process!` method occurred in a
transaction. Due to the fact that the payment was been created in the
initialization the entire action was not atomic.

Moving the transaction up to the controller where `.new(...).process!` is
should make the entire action atomic and prevent a new payment from
being created if the notification and redirect occur simultaneously.