Look up order with order number
fixes #41

Creating multiple orders in the same browser causes a 404 on the completion page if the order is already completed by an authorization notification. This happens because the order lookup is done by the guest token, and new order will have the same guest token if the cookie for it is already set.
Because of this the first order is found and then the system attempts to look up the payment by an identifier and because this is the wrong order it doesn't have this payment, causing a 404.