Use transaction with lock on order
Lock the order by row to prevent completing segments from creating a new
payment + payment source and attempting to complete the order.

Both sections do more or less the same thing so they really need to
happen in isolation as they branch based on whether or not the order is
completed.