Don't unbind submit events when selecting Adyen
This can cause issues in stores that have other event handlers bound to
the submit event on the form. Since Spree.disableSaveOnClick attaches an
anonymous function as an event handler, we can't simply unbind it.
Instead, we just re-enable the button if the Adyen validations fail and
the submission is cancelled.