Use the correct card to update payments
I ran into an issue where simply getting the last card returned from
Adyen was not the new one we just created. This caused the payment to be
updated with incorrect details. Since the new card is created earlier in
this method, we know for sure it will be the latest one created.