Ignore duplicate notifications from Adyen
Adyen has an "at-least-once" delivery system for notifications, meaning
that we may receive the same notification multiple times. Instead of
trying to create the notification again and rescuing from the exception,
simply check if it exists and respond with accept if it does.

From the Adyen docs:
"Duplicate notifications have the same corresponding values for the
eventCode and pspReference fields."