Add checks to payment flow for Adyen credit cards
The end goal is to treat all Adyen payments (HPP or credit cards) the
same during the checkout flow. We can't do that immediately because
notifications aren't associated with credit card payments correctly, so
when a payment is transitioned to `processing`, it never gets
transitioned out of it because it doesn't receive the success
notification. When this is done, we should stop returning false.