Refactor credit card payments
This includes a number of major changes to how credit cards work in the
gem:
- notifications are now associated with credit card payments
- credit card payments are transitioned based on notifications like HPP
- support for payment profiles was removed
- authorisation happens in the payment step
- auto_capture is properly supported on the cc payment method

The reason to auth in the payment step is that Adyen did not like that
we were doing $0 auths to "create profiles". Since this is the only step
we have access to the encrypted card data, we have to perform the auth
here. The only alternatives are to change the checkout flow or persist
the encrypted card data (neither of which are desirable).

Payment profile support was removed because Adyen doesn't truly support
payment profiles. They support the notion of recurring payment
contracts, which require a first payment transaction to store the
customer's details. This is why there was initially a $0 authorization
that passed in the flag to create a recurring contract.

I had to override #purchase! for credit cards to only call #capture!
instead, since we'll have already authorized the payment at that point.

NOTE: The biggest concern with this implementation is that if a customer
abandons checkout in the confirm step, there will be an outstanding
authorization on the card. Anyone using this functionality should
regularly cancel old authorizations.