Create recurring contracts for Adyen credit cards
This is largely the logic that happened in #create_profile before. We
still want to create a recurring contract with Adyen so that we store
the user's credit card info in case we need to use it to modify the
payment later.

Although payment_profiles_supported is still false for Adyen credit
cards, default Solidus will still list the cards on the payment step to
allow users to select them. This maintains support for that by calling
authorise_recurring_payment using the recurring contract ID instead when
we do the authorisation.