Migrate remaining payment requests to REST API
For the various calls there are options that we may want to provide that
aren't easily supported by the Adyen gem. One example is for the
openinvoice payment method, we need to provide invoice line data along
with partial captures and refunds.

Since the SOAP API implementation is just a collection of XML strings
concatenated together, it's very hard to customize this to provide
additional data. Using the REST API, we just provide a hash of
attributes with the payment request where we can stick whatever data we
need to.