Pass full gateway options to credit!
All other gateway actions receive the full gateway options, there's no
reason we shouldn't provide it here. The currency is included in the
gateway options by default.

This is also done so that we can use the gateway options to provide
additional data for Adyen. One example is to provide invoice line
information when performing partial refunds for invoice type payments.