Provide class to generate invoice request params
To create invoice payments Adyen requires that we provide invoice lines,
which correspond directly to line items. One challenge with this is that
we store VAT amounts on a per line item basis, but they want the VAT
amount per unit within a line item. We have to recompute the VAT amount
as if the quantity was 1 to avoid rounding errors.