Authorise Ratepay payments in after_create
In order to process Ratepay payments, we require the user's date of
birth. I'm setting the date of birth fields on the source as an
attr_accessor, meaning they get assigned after the `payment` checkout
step. Similar to credit card data, this is the only time we'll have
access to it, so we need to perform the authorization in an
after_create.

Note: If users go back in the checkout flow this will leave open
authorizations which should be cancelled either manually or
programatically.