Treat Ratepay payment processing like credit cards
We authorize both in the payment step, meaning that when we get to the
confirm page, the payment is already auth'd. The Solidus payment
processing flow will automatically call #authorize! for payment methods
that have auto_capture disabled and will call #purchase! for payment
methods that have it enabled. Since we're already auth'd, purchase
should just complete the capture for these payments.