Allow users to provide a merchant account per store
For applications with multiple stores, it's possible that users may want
transactions for different stores to be processed by different merchant
accounts in Adyen. A typical use case would be that the user has stores
for different countries and they have different bank accounts for each
country. Usually Adyen will link the bank accounts to different merchant
accounts and leave it up to us to pass the correct merchant account with
the payment request.

We don't always have access to the store when we're creating payment
requests, so I needed a way to look it up given the limited information
we have. In some cases, such as the `cancel` gateway action, the only
information we have is the psp reference for the payment. The
AccountLocator module provides several methods to look up the merchant
account when we don't have access to the store directly.