Allow Capybara driver to be changed by metadata
Sometimes it's easier to use Selenium in development so that we can
actually see what's going on when running our specs. This allows us to
just add `driver: :chrome` to the metadata of our examples to switch the
driver temporarily. With nothing specified it will default to using
poltergeist like we currently are.