Always set payments' request env in checkout controller
Normally solidus will set `payment.request_env` in
`Spree::PaymentCreate`. The `request_env` attribute just uses an
`attr_accessor`, so it's not persisted and will only be available after
the `payment` checkout step when the payment is create.

For 3DS payments with Adyen, we need to pass the HTTP_USER_AGENT and
HTTP_ACCEPT headers with the payment request, so we need access to them
after the `confirm` step. This will always set the request_env on any
created payments and then pass those two headers through with the
gateway options so we can use them in payment requests.