Add RedirectResponse model for 3DS payments
When dealing with 3DS payments with Adyen, we don't know whether a card
supports 3DS or not until we attempt to make the authorize call. When we
make the call one of two things can happen:
- the card does not support 3DS and a normal authorization happens
- the card supports 3DS and we get a URL to redirect to

This makes following the usual Solidus checkout flow awkward, because we
don't know what to do until after we make the authorize call to Adyen,
which happens in payment processing when transitioning from confirm ->
complete. If we get a redirect response back, we somehow need to
redirect before the order completes. Due to the order updater reloading
the order, we'll only have access to the response if we persist it,
which is what this model is intended to help with.