Backport process_payments_with from Solidus 1.4
There is an issue with the `process_payments_with` method where it will
raise a GatewayError if there are no unprocessed payments when
transitioning from confirm -> complete. It's possible with some of our
payment methods that we already have an authorized payment at that point
and there is no need for any further processing.

This issue was resolved in Solidus 1.4 in the following PR:
https://github.com/solidusio/solidus/pull/1361