Convert card holder name to UTF-8
The response we get back from Adyen when we look up payment profiles
says it should be in UTF-8, but what we really get back appears to be
ISO 8859-1.

This causes problems when dealing with names with non-ASCII characters
such as: ä, ü, etc.

These characters are represented as invalid utf-8 byte sequences and
will cause saving them to the database to fail.